+++ oykus/frontend/src/App.jsx
@@ -1,12 +1,12 @@
 import { useRouter } from "@/services/router";
-import Loading from "@/pages/Loading";
+import { OkpLoading } from "@/components/feedback";
 import Error404 from "@/pages/errors/404";
 
 export default function App() {
   const { View, props, params, isLoading } = useRouter();
 
   if (isLoading) {
-    return <Loading />;
+    return <OkpLoading />;
   }
 
   if (!View) {