+++ oykus/frontend/src/types/components/feedback/loading.ts
@@ -1,0 +1,4 @@
+export interface LoadingProps {
+  variant?: "squares" | "layers" | "circles";
+  py?: number;
+}